pdf/pdf2img_withson/requirements.txt
fitz
pymupdf
matplotlib>=3.2.2
numpy
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0
protobuf<4.21.3
